fix: reject wrapped u32 indexes on 32-bit targets

+++ src/core/compiler/ir/build.go
@@ -205,7 +205,7 @@ func BuildFunc(m *wasm.Module, localFuncIdx int) (*Func, error) {
 }
 
 func (b *Builder) buildFunc(localIdx uint32) (*Func, error) {
-	if int(localIdx) >= len(b.m.FuncTypes) || int(localIdx) >= len(b.m.Code) {
+	if uint(localIdx) >= uint(len(b.m.FuncTypes)) || uint(localIdx) >= uint(len(b.m.Code)) {
 		return nil, fmt.Errorf("ir: local function index %d out of range", localIdx)
 	}
 	typeIdx := b.m.FuncTypes[localIdx].Index
@@ -1353,7 +1353,7 @@ func (b *Builder) lowerFC() error {
 		if err != nil {
 			return err
 		}
-		if int(data) >= len(b.m.Data) {
+		if uint(data) >= uint(len(b.m.Data)) {
 			return fmt.Errorf("unknown data segment %d", data)
 		}
 		mem, err := b.readZeroMemoryImmediate()
@@ -1385,7 +1385,7 @@ func (b *Builder) lowerFC() error {
 		if err != nil {
 			return err
 		}
-		if int(data) >= len(b.m.Data) {
+		if uint(data) >= uint(len(b.m.Data)) {
 			return fmt.Errorf("unknown data segment %d", data)
 		}
 		if b.reachable {
@@ -1611,7 +1611,7 @@ func (b *Builder) setReturn(args []ValueID) {
 }
 func (b *Builder) setTrap() { b.fn.Blocks[b.cur].Term = Term{Kind: TermTrap} }
 func (b *Builder) labelAt(depth uint32) (label, error) {
-	if int(depth) >= len(b.labels) {
+	if uint(depth) >= uint(len(b.labels)) {
 		return label{}, fmt.Errorf("unknown label depth %d", depth)
 	}
 	return b.labels[len(b.labels)-1-int(depth)], nil
@@ -1675,7 +1675,7 @@ func (b *Builder) funcType(fi uint32) (*wasm.FuncType, error) {
 	if err != nil {
 		return nil, err
 	}
-	if int(ti) >= len(b.out.Types) {
+	if uint(ti) >= uint(len(b.out.Types)) {
 		return nil, fmt.Errorf("unknown type %d", ti)
 	}
 	if !irTypeIsFunc(b.out, ti) {
@@ -1690,13 +1690,13 @@ func (b *Builder) funcType(fi uint32) (*wasm.FuncType, error) {
 func (b *Builder) funcTypeIndex(fi uint32) (uint32, error) {
 	// Function metadata is flattened once when the builder is created. Use it for
 	// O(1) call validation instead of re-scanning imports for every call opcode.
-	if int(fi) >= len(b.out.FuncTypes) {
+	if uint(fi) >= uint(len(b.out.FuncTypes)) {
 		return 0, fmt.Errorf("unknown function %d", fi)
 	}
 	return b.out.FuncTypes[fi], nil
 }
 func (b *Builder) globalType(x uint32) (wasm.GlobalType, error) {
-	if int(x) >= len(b.out.Globals) {
+	if uint(x) >= uint(len(b.out.Globals)) {
 		return wasm.GlobalType{}, fmt.Errorf("unknown global %d", x)
 	}
 	gt := b.out.Globals[x]
@@ -1709,19 +1709,19 @@ func (b *Builder) memoryType(x uint32) (wasm.MemType, error) {
 	if x != 0 {
 		return wasm.MemType{}, fmt.Errorf("multi-memory unsupported: memory index %d", x)
 	}
-	if int(x) >= len(b.out.Memories) {
+	if uint(x) >= uint(len(b.out.Memories)) {
 		return wasm.MemType{}, fmt.Errorf("unknown memory %d", x)
 	}
 	return b.out.Memories[x], nil
 }
 func (b *Builder) tableType(x uint32) (wasm.TableType, error) {
-	if int(x) >= len(b.out.Tables) {
+	if uint(x) >= uint(len(b.out.Tables)) {
 		return wasm.TableType{}, fmt.Errorf("unknown table %d", x)
 	}
 	return b.out.Tables[x], nil
 }
 func typeOf(f *Func, v ValueID) wasm.ValType {
-	if v == InvalidValue || int(v) >= len(f.Values) {
+	if v == InvalidValue || uint(v) >= uint(len(f.Values)) {
 		return wasm.ValType{}
 	}
 	return f.Values[v].Type
