+++ src/core/compiler/ir/verify.go
@@ -26,7 +26,7 @@ func VerifyModule(m *Module) error {
 		return err
 	}
 	for i := range m.FuncTypes {
-		if int(m.FuncTypes[i]) >= len(m.Types) {
+		if uint(m.FuncTypes[i]) >= uint(len(m.Types)) {
 			return fmt.Errorf("ir: function %d has unknown type %d", i, m.FuncTypes[i])
 		}
 		if !irTypeIsFunc(m, m.FuncTypes[i]) {
@@ -68,7 +68,7 @@ func verifyCanonicalTypeIDs(m *Module) error {
 			continue
 		}
 		canon := m.CanonicalTypeIDs[i]
-		if int(canon) >= len(m.Types) {
+		if uint(canon) >= uint(len(m.Types)) {
 			return fmt.Errorf("ir: canonical type id for type %d out of range: %d", i, canon)
 		}
 		if !irTypeIsFunc(m, canon) {
@@ -92,7 +92,7 @@ func verifyCanonicalTypeIDs(m *Module) error {
 }
 
 func verifyModuleFuncHeaders(m *Module) error {
-	if int(m.ImportedFuncCount) > len(m.FuncTypes) {
+	if uint(m.ImportedFuncCount) > uint(len(m.FuncTypes)) {
 		return fmt.Errorf("ir: imported function count %d exceeds function type count %d", m.ImportedFuncCount, len(m.FuncTypes))
 	}
 	wantLocalFuncs := len(m.FuncTypes) - int(m.ImportedFuncCount)
@@ -147,7 +147,7 @@ func verifyFunc(f *Func, m *Module) error {
 	if f == nil {
 		return fmt.Errorf("ir: nil func")
 	}
-	if int(f.Entry) >= len(f.Blocks) {
+	if uint(f.Entry) >= uint(len(f.Blocks)) {
 		return fmt.Errorf("entry block %d out of range", f.Entry)
 	}
 	if err := verifyLocalLayout(f); err != nil {
@@ -160,11 +160,11 @@ func verifyFunc(f *Func, m *Module) error {
 		}
 		switch v.DefKind {
 		case ValueDefBlockParam:
-			if int(v.Def) >= len(f.Blocks) {
+			if uint(v.Def) >= uint(len(f.Blocks)) {
 				return fmt.Errorf("value %d has invalid block def %d", i, v.Def)
 			}
 		case ValueDefInst:
-			if int(v.Def) >= len(f.Insts) {
+			if uint(v.Def) >= uint(len(f.Insts)) {
 				return fmt.Errorf("value %d has invalid inst def %d", i, v.Def)
 			}
 		case ValueDefPoison:
@@ -744,7 +744,7 @@ func verifyGlobalAccess(m *Module, id InstID, in *Inst, got wasm.ValType) error 
 		return nil
 	}
 	idx := uint32(in.Aux)
-	if int(idx) >= len(m.Globals) {
+	if uint(idx) >= uint(len(m.Globals)) {
 		return fmt.Errorf("inst %d global index %d out of range", id, idx)
 	}
 	want := globalTypeValue(m.Globals[idx])
@@ -863,7 +863,7 @@ func verifyMemoryIndex(m *Module, id InstID, idx uint32) error {
 	if idx != 0 {
 		return fmt.Errorf("inst %d multi-memory unsupported: memory index %d", id, idx)
 	}
-	if m != nil && int(idx) >= len(m.Memories) {
+	if m != nil && uint(idx) >= uint(len(m.Memories)) {
 		return fmt.Errorf("inst %d memory index %d out of range", id, idx)
 	}
 	return nil
@@ -919,7 +919,7 @@ func verifyCall(m *Module, id InstID, in *Inst, argc, resc int, argt, rest func(
 	if in.Op == OpCallIndirect {
 		typeIdx := callIndirectType(in.Aux)
 		tableIdx := callIndirectTable(in.Aux)
-		if int(typeIdx) >= len(m.Types) {
+		if uint(typeIdx) >= uint(len(m.Types)) {
 			return fmt.Errorf("inst %d call_indirect type %d out of range", id, typeIdx)
 		}
 		if !irTypeIsFunc(m, typeIdx) {
@@ -928,7 +928,7 @@ func verifyCall(m *Module, id InstID, in *Inst, argc, resc int, argt, rest func(
 		if got, want := uint32(in.Aux2), irCanonicalTypeID(m, typeIdx); got != want {
 			return fmt.Errorf("inst %d call_indirect canonical type id %d, want %d", id, got, want)
 		}
-		if int(tableIdx) >= len(m.Tables) {
+		if uint(tableIdx) >= uint(len(m.Tables)) {
 			return fmt.Errorf("inst %d call_indirect table %d out of range", id, tableIdx)
 		}
 		if !irIsFuncRefTableType(m, tableRefType(m.Tables[tableIdx])) {
@@ -955,7 +955,7 @@ func verifyCall(m *Module, id InstID, in *Inst, argc, resc int, argt, rest func(
 		return nil
 	}
 	fi := uint32(in.Aux)
-	if int(fi) >= len(m.FuncTypes) {
+	if uint(fi) >= uint(len(m.FuncTypes)) {
 		return fmt.Errorf("inst %d call function %d out of range", id, fi)
 	}
 	if in.Op == OpCallImport && fi >= m.ImportedFuncCount {
@@ -965,7 +965,7 @@ func verifyCall(m *Module, id InstID, in *Inst, argc, resc int, argt, rest func(
 		return fmt.Errorf("inst %d call function %d is imported", id, fi)
 	}
 	typeIdx := m.FuncTypes[fi]
-	if int(typeIdx) >= len(m.Types) {
+	if uint(typeIdx) >= uint(len(m.Types)) {
 		return fmt.Errorf("inst %d call function %d has unknown type %d", id, fi, typeIdx)
 	}
 	if !irTypeIsFunc(m, typeIdx) {
@@ -994,7 +994,7 @@ func irIsFuncRefTableType(m *Module, rt wasm.RefType) bool {
 	case wasm.HeapAbs:
 		return heap.Abs() == wasm.HeapFunc || heap.Abs() == wasm.HeapNoFunc
 	case wasm.HeapTypeIndex:
-		if m == nil || heap.Type().Rec || int(heap.Type().Index) >= len(m.Types) {
+		if m == nil || heap.Type().Rec || uint(heap.Type().Index) >= uint(len(m.Types)) {
 			return false
 		}
 		return irTypeIsFunc(m, heap.Type().Index)
@@ -1067,7 +1067,7 @@ func verifyEdges(f *Func, bid BlockID, r Range) error {
 	}
 	for ei := r.Start; ei < end; ei++ {
 		e := f.Edges[ei]
-		if int(e.To) >= len(f.Blocks) {
+		if uint(e.To) >= uint(len(f.Blocks)) {
 			return fmt.Errorf("block %d edge %d target %d out of range", bid, ei, e.To)
 		}
 		if _, err := verifyValueRange(f, e.Args, "edge args"); err != nil {
@@ -1157,7 +1157,7 @@ func verifyDominance(f *Func) error {
 			}
 			return nil
 		}
-		if int(defBlock) >= len(reachable) || !reachable[defBlock] || !dominatesInterval(domPre, domEnd, defBlock, use) {
+		if uint(defBlock) >= uint(len(reachable)) || !reachable[defBlock] || !dominatesInterval(domPre, domEnd, defBlock, use) {
 			return fmt.Errorf("%s value %d from b%d does not dominate b%d", what, v, defBlock, use)
 		}
 		return nil
@@ -1219,7 +1219,7 @@ func branchEdges(t *Term) (Range, bool) {
 
 func reversePostorder(entry BlockID, succs [][]BlockID) ([]bool, []BlockID) {
 	reachable := make([]bool, len(succs))
-	if int(entry) >= len(succs) {
+	if uint(entry) >= uint(len(succs)) {
 		return reachable, nil
 	}
 	type frame struct {
@@ -1259,7 +1259,7 @@ func computeIDoms(entry BlockID, preds [][]BlockID, reachable []bool, rpo []Bloc
 	for i, b := range rpo {
 		order[b] = int32(i)
 	}
-	if int(entry) >= len(idom) {
+	if uint(entry) >= uint(len(idom)) {
 		return idom, order
 	}
 	idom[entry] = entry
@@ -1312,7 +1312,7 @@ func dominanceIntervals(entry BlockID, idom []BlockID, reachable []bool) ([]int3
 	for i := range pre {
 		pre[i], end[i] = -1, -1
 	}
-	if int(entry) >= len(idom) || !reachable[entry] {
+	if uint(entry) >= uint(len(idom)) || !reachable[entry] {
 		return pre, end
 	}
 	children := make([][]BlockID, len(idom))
@@ -1350,7 +1350,7 @@ func dominanceIntervals(entry BlockID, idom []BlockID, reachable []bool) ([]int3
 }
 
 func dominatesInterval(pre, end []int32, a, b BlockID) bool {
-	if int(a) >= len(pre) || int(b) >= len(pre) || pre[a] < 0 || pre[b] < 0 {
+	if uint(a) >= uint(len(pre)) || uint(b) >= uint(len(pre)) || pre[a] < 0 || pre[b] < 0 {
 		return false
 	}
 	return pre[a] <= pre[b] && pre[b] < end[a]
@@ -1379,7 +1379,7 @@ func verifyRange(r Range, total int, what string) (uint32, error) {
 	return r.Start + r.Len, nil
 }
 func verifyValue(f *Func, v ValueID, what string) error {
-	if v == InvalidValue || int(v) >= len(f.Values) {
+	if v == InvalidValue || uint(v) >= uint(len(f.Values)) {
 		return fmt.Errorf("%s invalid value %d", what, v)
 	}
 	return nil
